Reorganize linker related code.
