+++ src/tools/wix/Link/ConnectToFeature.cs
@@ -11,14 +11,9 @@
 // </summary>
 //-------------------------------------------------------------------------------------------------
 
-namespace WixToolset
+namespace WixToolset.Link
 {
-    using System;
-    using System.Diagnostics;
-    using System.Collections;
     using System.Collections.Specialized;
-    using System.Globalization;
-    using System.Xml;
     using WixToolset.Data;
 
     /// <summary>
