+++ src/tools/wix/Link/ConnectToFeatureCollection.cs
@@ -11,7 +11,7 @@
 // </summary>
 //-------------------------------------------------------------------------------------------------
 
-namespace WixToolset
+namespace WixToolset.Link
 {
     using System;
     using System.Collections;
